producer/requirements.txt
Faker==26.0.0
kafka-python==2.0.2
python-dateutil==2.9.0.post0
six==1.16.0
python-dotenv==1.0.1